Gather logs from failed cluster bootstrapping.
Adds a PVC for each new install job. If the job is successful, no logs
are gathered and the PVC is deleted immediately.

If an error occurs during install we will try to gather logs from the
bootstrap node and copy them to the PVC volume. Any error in this
process is not fatal. The PVC will be presered for 7 days, after which
the next reconcile will delete it and the logs are gone. Each install
attempt will copy it's timestamped tarball into the same PVC, so the
logs from all attempts are preserved.

The volume will store up to 1GB in logs at which point log gathering
will start to fail. This should be extremly unlikely to occur, non-fatal
if it does, and at that point we should certainly have enough attempts
to discern what is going on.

A followup effort will try to gather logs from the cluster if we get
past bootstrap, teardown the bootstrap node, but fail during the final
phase of install. (operators etc)

Feature is disabled by default and can be enabled via hiveconfig.