Restart deprovision pods if credentials secret changes.
Previously the uninstall pod would never restart if it launched with
since deleted or bad creds, and the user updated the credentials secret.

This change switches to mounting the creds as a secret instead of
passing through env vars. The hiveutil wrapper code sets them as env
vars and establishes a file watch on the directory where the secret is
mounted. On any change or watch error we exit to restart the pod.

It seems to usually take about 1 minute for the secret changes to
surface in the pod.